test(server): fold the gatewaycompat wiring tests into main_test.go
The naming convention allows a _test.go only under the name of the
module it tests, and the wiring under test lives in main.go. Content
unchanged, verified under -shuffle=on.
